+++ src/utils/adapter/addressToString/index.ts
@@ -1,0 +1,4 @@
+import addressToString from "./addressToString";
+
+export default addressToString;
+export * from "./addressToString";